# TST: Dynamically use doctest_namespace only if running the doctest