@mroeschke 2022-01-28
cc @jreback green and should speed up the CI tests