@mroeschke 2022-01-28
Hmm might be a good idea but not critical since this is not a CI-configuration change per se.