@jreback 2022-01-28
@meeseeksdev backport 1.4.x